refactor code in admin, add chatbox viewmodel

+++ app/src/main/java/com/example/foodapp/data/dto/request/ChatKnowledgeEntryRequest.kt
@@ -3,5 +3,5 @@ package com.example.foodapp.data.dto.request
 data class ChatKnowledgeEntryRequest(
     val title: String,
     val content: String,
-    val intentTypeId: Int
+    val intentTypeId: Long
 )
